+++ src/main/kotlin/com/prasetia/erp/model/corrective/CorrectiveAdvanceInvoice.kt
@@ -7,10 +7,10 @@ import javax.persistence.Id
 class CorrectiveAdvanceInvoice(
         @Id
         val id:Long,
-        val nilai_invoice:Long,
+        val nilai_invoice:Double,
         val invoice_state:String,
         val no_inv:String,
         val year_project:String
 ){
-    constructor(): this(0, 0, "", "", "")
+    constructor(): this(0, 0.0, "", "", "")
 }
